+++ app/src/main/java/com/adammcneilly/pocketleague/ui/OnePane.kt
@@ -7,8 +7,8 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.saveable.SaveableStateHolder
 import androidx.compose.ui.Modifier
-import com.adammcneilly.pocketleague.composables.Level1BottomBar
-import com.adammcneilly.pocketleague.composables.TopBar
+import com.adammcneilly.pocketleague.composables.bars.Level1BottomBar
+import com.adammcneilly.pocketleague.composables.bars.TopBar
 import com.adammcneilly.pocketleague.shared.screens.Navigation
 
 /**
