+++ app/src/main/java/com/adammcneilly/pocketleague/ui/TwoPane.kt
@@ -13,8 +13,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.saveable.SaveableStateHolder
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.adammcneilly.pocketleague.composables.Level1NavigationRail
-import com.adammcneilly.pocketleague.composables.TopBar
+import com.adammcneilly.pocketleague.composables.bars.Level1NavigationRail
+import com.adammcneilly.pocketleague.composables.bars.TopBar
 import com.adammcneilly.pocketleague.shared.screens.Navigation
 import com.adammcneilly.pocketleague.shared.screens.ScreenIdentifier
