+++ composables/src/commonMain/kotlin/com/adammcneilly/pocketleague/composables/bars/Level1NavigationRail.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.pocketleague.composables
+package com.adammcneilly.pocketleague.composables.bars
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.BarChart
